feat: applied clean code
